+++ packages/ui-react/src/components/Deposit/DepositInfo/index.tsx
@@ -25,6 +25,7 @@ export interface DepositInfoProps {
   extraNotes?: string[];
   serviceFee?: string;
   serviceFeeUsd?: string;
+  threshold?: string;
   modalContainer?: CommonModalProps['getContainer'];
   componentStyle?: ComponentStyle;
   customDescriptionNode?: React.ReactNode;
@@ -41,6 +42,7 @@ export default function DepositInfo({
   extraNotes,
   serviceFee,
   serviceFeeUsd,
+  threshold,
   modalContainer,
   componentStyle = ComponentStyle.Web,
   customDescriptionNode,
@@ -79,6 +81,20 @@ export default function DepositInfo({
     );
   }, [customDescriptionNode, extraNotes]);
 
+  const serviceFeeTip = useMemo(() => {
+    return (
+      <div>
+        <div>{SERVICE_FEE_TIP}</div>
+        <div>{`• Deposit amount ≥ ${threshold} ${formatSymbolDisplay(depositTokenSymbol)}: No service fee`}</div>
+        <div>
+          {`• Deposit amount < ${threshold} ${formatSymbolDisplay(
+            depositTokenSymbol,
+          )}: Max service fee ${serviceFee} ${formatSymbolDisplay(depositTokenSymbol)}`}
+        </div>
+      </div>
+    );
+  }, [depositTokenSymbol, serviceFee, threshold]);
+
   return (
     <div className={clsx('etransfer-ui-flex-column etransfer-ui-deposit-info', className)}>
       {!!serviceFee && serviceFee !== '0' && (
@@ -86,7 +102,7 @@ export default function DepositInfo({
           <div className={clsx('etransfer-ui-flex-row-center etransfer-ui-gap-4', 'info-title')}>
             {SERVICE_FEE}
             <CommonTip
-              tip={SERVICE_FEE_TIP}
+              tip={serviceFeeTip}
               className={'service-fee-tip'}
               modalTitle={NOTICE}
               icon={<CommonSvg type="questionMark16" />}
@@ -95,10 +111,10 @@ export default function DepositInfo({
           </div>
           <div className={clsx('etransfer-ui-flex-1')}>
             <div className={clsx('etransfer-ui-text-right', 'info-value')}>
-              {serviceFee} {formatSymbolDisplay(depositTokenSymbol)}
+              {`0~${serviceFee}`} {formatSymbolDisplay(depositTokenSymbol)}
             </div>
             <div className={clsx('etransfer-ui-text-right', 'info-exhibit')}>
-              {valueFixed2LessThanMin(serviceFeeUsd, '$ ')}
+              {`$ 0~${valueFixed2LessThanMin(serviceFeeUsd, '')}`}
             </div>
           </div>
         </div>
@@ -111,7 +127,7 @@ export default function DepositInfo({
               {minimumDeposit} {formatSymbolDisplay(depositTokenSymbol)}
             </div>
             <div className={clsx('etransfer-ui-text-right', 'info-exhibit')}>
-              {valueFixed2LessThanMin(minAmountUsd, '$ ')}
+              {minAmountUsd === '0' ? '$ 0' : valueFixed2LessThanMin(minAmountUsd, '$ ')}
             </div>
           </div>
         </div>